requirements.txt
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0
transformers>=4.51.1
pillow>=10.0.0
python-Levenshtein>=0.23.0
PyMuPDF>=1.23.0
tqdm>=4.66.0
safetensors>=0.4.0
flash-attn>=2.7.3
